+++ cfg.py
@@ -12,7 +12,7 @@ show_debug_err = False
 no_sending_to_server = False
 
 # write storing to server content to file
-to_server_to_file = False
+to_server_to_file = True
 
 # write loaded from server content to file
-from_server_to_file = False
+from_server_to_file = True